fix(vercel): puppeteer attempt
😫
