+++ package.json
@@ -9,10 +9,12 @@
   },
   "dependencies": {
     "@monaco-editor/react": "^4.4.6",
+    "chrome-aws-lambda": "^10.1.0",
     "lodash": "^4.17.21",
     "next": "12",
     "next-seo": "^5.5.0",
     "playwright-core": "^1.27.0",
+    "puppeteer-core": "^18.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sharp": "^0.31.1",
